+++ packages/gatsby-plugin-image/src/resolver-utils.ts
@@ -168,7 +168,7 @@ export function getGatsbyImageFieldConfig<TSource, TContext>(
             BLURRED: a blurred, low resolution image, encoded as a base64 data URI (default)
             DOMINANT_COLOR: a solid color, calculated from the dominant color of the image.
             TRACED_SVG: a low-resolution traced SVG of the image.
-            NONE: no placeholder. Set "background" to use a fixed background color.`,
+            NONE: no placeholder. Set the argument "backgroundColor" to use a fixed background color.`,
       },
       formats: {
         type: GraphQLList(ImageFormatType),
@@ -176,9 +176,9 @@ export function getGatsbyImageFieldConfig<TSource, TContext>(
             The image formats to generate. Valid values are AUTO (meaning the same format as the source image), JPG, PNG, WEBP and AVIF.
             The default value is [AUTO, WEBP], and you should rarely need to change this. Take care if you specify JPG or PNG when you do
             not know the formats of the source images, as this could lead to unwanted results such as converting JPEGs to PNGs. Specifying
-            both PNG and JPG is not supported and will be ignored. AVIF support is currently experimental.
+            both PNG and JPG is not supported and will be ignored. 
         `,
-        defaultValue: [`auto`, `webp`],
+        defaultValue: [``, `webp`],
       },
       outputPixelDensities: {
         type: GraphQLList(GraphQLFloat),