+++ packages/gatsby/src/commands/build-html.ts
@@ -7,6 +7,7 @@ import webpack from "webpack"
 import * as path from "path"
 
 import { emitter, store } from "../redux"
+import { IWebpackWatchingPauseResume } from "../utils/start-server"
 import webpackConfig from "../utils/webpack.config"
 import { structureWebpackErrors } from "../utils/webpack-error-utils"
 import * as buildUtils from "./build-utils"
@@ -18,16 +19,13 @@ import { PackageJson } from "../.."
 type IActivity = any // TODO
 type IWorkerPool = any // TODO
 
-export interface IWebpackWatchingPauseResume extends webpack.Watching {
-  suspend: () => void
-  resume: () => void
-}
-
 export interface IBuildArgs extends IProgram {
   directory: string
   sitePackageJson: PackageJson
   prefixPaths: boolean
   noUglify: boolean
+  logPages: boolean
+  writeToFile: boolean
   profile: boolean
   graphqlTracing: boolean
   openTracingConfigFile: string
@@ -37,11 +35,11 @@ export interface IBuildArgs extends IProgram {
 let devssrWebpackCompiler: webpack.Compiler
 let devssrWebpackWatcher: IWebpackWatchingPauseResume
 let needToRecompileSSRBundle = true
-export const getDevSSRWebpack = (): Record<
-  IWebpackWatchingPauseResume,
-  webpack.Compiler,
-  needToRecompileSSRBundle
-> => {
+export const getDevSSRWebpack = (): {
+  devssrWebpackWatcher: IWebpackWatchingPauseResume
+  devssrWebpackCompiler: webpack.Compiler
+  needToRecompileSSRBundle: boolean
+} => {
   if (process.env.gatsby_executing_command !== `develop`) {
     throw new Error(`This function can only be called in development`)
   }
@@ -104,7 +102,7 @@ const runWebpack = (
             return resolve(stats)
           }
         }
-      )
+      ) as IWebpackWatchingPauseResume
     }
   })
 
@@ -154,6 +152,10 @@ export const deleteRenderer = async (rendererPath: string): Promise<void> => {
   }
 }
 
+export interface IRenderHtmlResult {
+  unsafeBuiltinsUsageByPagePath: Record<string, Array<string>>
+}
+
 const renderHTMLQueue = async (
   workerPool: IWorkerPool,
   activity: IActivity,
@@ -178,25 +180,71 @@ const renderHTMLQueue = async (
       ? workerPool.renderHTMLProd
       : workerPool.renderHTMLDev
 
-  await Bluebird.map(segments, async pageSegment => {
-    await renderHTML({
-      envVars,
-      htmlComponentRendererPath,
-      paths: pageSegment,
-      sessionId,
-    })
+  const uniqueUnsafeBuiltinUsedStacks = new Set<string>()
 
-    if (stage === `build-html`) {
+  try {
+    await Bluebird.map(segments, async pageSegment => {
+      const htmlRenderMeta: IRenderHtmlResult = await renderHTML({
+        envVars,
+        htmlComponentRendererPath,
+        paths: pageSegment,
+        sessionId,
+      })
+
+      if (stage === `build-html`) {
+        store.dispatch({
+          type: `HTML_GENERATED`,
+          payload: pageSegment,
+        })
+
+        for (const [_pagePath, arrayOfUsages] of Object.entries(
+          htmlRenderMeta.unsafeBuiltinsUsageByPagePath
+        )) {
+          for (const unsafeUsageStack of arrayOfUsages) {
+            uniqueUnsafeBuiltinUsedStacks.add(unsafeUsageStack)
+          }
+        }
+      }
+
+      if (activity && activity.tick) {
+        activity.tick(pageSegment.length)
+      }
+    })
+  } catch (e) {
+    if (e?.context?.unsafeBuiltinsUsageByPagePath) {
+      for (const [_pagePath, arrayOfUsages] of Object.entries(
+        e.context.unsafeBuiltinsUsageByPagePath
+      )) {
+        // @ts-ignore TS doesn't know arrayOfUsages is Iterable
+        for (const unsafeUsageStack of arrayOfUsages) {
+          uniqueUnsafeBuiltinUsedStacks.add(unsafeUsageStack)
+        }
+      }
+    }
+    throw e
+  } finally {
+    if (uniqueUnsafeBuiltinUsedStacks.size > 0) {
+      console.warn(
+        `Unsafe builtin method was used, future builds will need to rebuild all pages`
+      )
       store.dispatch({
-        type: `HTML_GENERATED`,
-        payload: pageSegment,
+        type: `SSR_USED_UNSAFE_BUILTIN`,
       })
     }
 
-    if (activity && activity.tick) {
-      activity.tick(pageSegment.length)
+    for (const unsafeBuiltinUsedStack of uniqueUnsafeBuiltinUsedStacks) {
+      const prettyError = await createErrorFromString(
+        unsafeBuiltinUsedStack,
+        `${htmlComponentRendererPath}.map`
+      )
+
+      const warningMessage = `${prettyError.stack}${
+        prettyError.codeFrame ? `\n\n${prettyError.codeFrame}\n` : ``
+      }`
+
+      reporter.warn(warningMessage)
     }
-  })
+  }
 }
 
 class BuildHTMLError extends Error {
@@ -271,13 +319,9 @@ export async function buildHTMLPagesAndDeleteStaleArtifacts({
 }> {
   buildUtils.markHtmlDirtyIfResultOfUsedStaticQueryChanged()
 
-  const { toRegenerate, toDelete } = process.env
-    .GATSBY_EXPERIMENTAL_PAGE_BUILD_ON_DATA_CHANGES
-    ? buildUtils.calcDirtyHtmlFiles(store.getState())
-    : {
-        toRegenerate: [...store.getState().pages.keys()],
-        toDelete: [],
-      }
+  const { toRegenerate, toDelete } = buildUtils.calcDirtyHtmlFiles(
+    store.getState()
+  )
 
   if (toRegenerate.length > 0) {
     const buildHTMLActivityProgress = reporter.createProgress(