+++ packages/gatsby-transformer-sqip/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-next.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.7...gatsby-transformer-sqip@3.0.0-next.8) (2021-02-22)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+# [3.0.0-next.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.6...gatsby-transformer-sqip@3.0.0-next.7) (2021-02-19)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+# [3.0.0-next.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.5...gatsby-transformer-sqip@3.0.0-next.6) (2021-02-19)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+# [3.0.0-next.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.4...gatsby-transformer-sqip@3.0.0-next.5) (2021-02-18)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
+# [3.0.0-next.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.3...gatsby-transformer-sqip@3.0.0-next.4) (2021-02-17)
+
+**Note:** Version bump only for package gatsby-transformer-sqip
+
 # [3.0.0-next.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-sqip@3.0.0-next.2...gatsby-transformer-sqip@3.0.0-next.3) (2021-02-12)
 
 **Note:** Version bump only for package gatsby-transformer-sqip